style: remove uint256  alias

+++ src/interfaces/IComp.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.8.15;
 
 interface IComp {
-    function getPriorVotes(address account, uint blockNumber) external view returns (uint96);
+    function getPriorVotes(address account, uint256 blockNumber) external view returns (uint96);
 }
